Changed Long to int in id of database and added api for deleting entitis by id

+++ SpringData/src/main/java/pl/krystian/Car.java
@@ -13,7 +13,7 @@ public class Car {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private int id;
 	
 	private String mark;
 	private String model;
@@ -22,10 +22,11 @@ public class Car {
 	public Car() {}
 	
 	
-	public Long getId() {
+	
+	public int getId() {
 		return id;
 	}
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 	public String getMark() {
